Serious security bug fixed

+++ app/commands/_reboot.php
@@ -1,4 +1,9 @@
 <?php
+session_start();
+
+if($_SESSION['username'] == ""){
+	die("You are not logged in");
+}
 	system('sudo reboot');
 ?>
